changed double to int and casted in move ()

+++ Assignment 1/Car.java
@@ -23,7 +23,7 @@ public abstract class Car extends Vehicle{
         this.enginePower = enginePower;
     }
 
-    public void setCurrentSpeed(double currentSpeed) {
+    public void setCurrentSpeed(int currentSpeed) {
         this.currentSpeed = currentSpeed;
     }
 
@@ -44,14 +44,15 @@ public abstract class Car extends Vehicle{
    @Override
    public void move(){
        if(getDirX()!= 0) {
-           setPosX(getPosX() + getDirX() * getCurrentSpeed());
+           setPosX((int) (getPosX() + getDirX() * getCurrentSpeed()));
        }else if(getDirY()!= 0) {
-           setPosY(getPosY() + getDirY() * getCurrentSpeed());
+           setPosY((int) (getPosY() + getDirY() * getCurrentSpeed()));
        }
    }
 
    @Override
    public void turnLeft() {
+
        if (getDirX() == -1 && getDirY() == 0) {
            setDirX(0);
            setDirY(-1);
